fix: workaround for 'thread already stopped' nvim issue
NeoVim gets upset about 'step' events for threads which are not stopped.
Normally the server would send an event for both CPU and copper threads,
even though only one of them is stopped. The error message from NeoVim,
'thread already stopped' seems to be misleading as the opposite is true.

Inversely, sending a breakpoint stop event for a stopped thread also
causes issues.